requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2019.3.9
chardet==3.0.4
confusable-homoglyphs==3.2.0
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.6
django-crispy-forms==1.7.2
django-import-export==1.2.0
django-tinymce4-lite==1.7.4
django-tinymce4-widget==4.0.0
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.8
isort==4.3.4
jdcal==1.4.1
jsmin==2.2.2
lazy-object-proxy==1.3.1
MarkupPy==1.14
mccabe==0.6.1
mysql-connector-python==8.0.18
odfpy==1.4.0
openpyxl==3.0.0
pep8==1.7.1
Pillow==5.4.1
protobuf==3.10.0
psycopg2==2.7.7
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.2.2
pytz==2017.2
PyYAML==5.1.2
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
tablib==0.14.0
typed-ast==1.1.1
urllib3==1.24.1
validate-email==1.3
virtualenv==16.7.9
whitenoise==4.1.2
wrapt==1.11.0
xlrd==1.2.0
xlwt==1.3.0
